drag the new HPET code into oxide because it's mostly shareable
This still won't actually work, because for some reason this work was done
such that the ACPI table is the unit of abstraction instead of the address of
the HPET's MMIO register block (which is the only thing we really need in that
table).  In order to commonise this, that abstraction needs to change, and the
i86pc implementation should look up the MMIO address itself and pass that into
an ACPI-less HPET initialisation routine.  Once that's done, the HPET code
becomes entirely generic and the oxide caller can use the same init entry
point, "finding" the HPET by virtue of reading the manual instead of reading a
superfluous ACPI table.  Since all oxide machines have a working PIT, the
temporary inability to use this new code does no harm.