initialize topology and train PCIe
This goes through and does a large amount. Importantly, it creates and
fills out topology information in the system, provides access to the
SMN and data fabric, and then performs a large chunk of initialization
(though this is incomplete). From there, we actually are able to use
board-specific knowledge to program the dxio subsystem. With this,
you're able to actually link up and train devices.