Iniitialize basic PCI support in the system
This, albeit with some hacks for pci_autoconfig, should bring up PCIe on
Gimlet and Ethanol-X. This does not bring up hotplug on Gimlet at this
time. In addition, we add the oxide platform pci_prd module, which does
resource discovery directly from the hardware.

As part of this reshuffle some of the files in the manifest and remove
the ISA PCI binding so we can make things work. There are some
additional entries in the ramdisk manifest to attempt to get NVMe, PCIe
going.